+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Expose supported theme hooks as public constant
+- Modify theme registry for only hooks that exist
 
 ### Security
 - Implement `TrustedCallbackInterface` for `#pre_render`