+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2020-04-21
 ### Added
 - Add tests
 
@@ -74,7 +76,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix coding standards violation
 - Fix typo in comment
 
-[Unreleased]: https://github.com/wongjn/lazy_image/compare/1.3.8...HEAD
+[Unreleased]: https://github.com/wongjn/lazy_image/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/wongjn/lazy_image/compare/1.3.8...v2.0.0
 [1.3.8]: https://github.com/wongjn/lazy_image/compare/1.3.7...1.3.8
 [1.3.7]: https://github.com/wongjn/lazy_image/compare/1.3.6...1.3.7
 [1.3.6]: https://github.com/wongjn/lazy_image/compare/1.3.5...1.3.6